# el-pasas/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "0.1",
    "title" : "eDocument API",
    "termsOfService" : ""
  },
  "host" : "el-pasas.azurewebsites.net",
  "basePath" : "/api",
...
  "security" : [ {
    "Token" : [ ]
  } ],
  "paths" : {
    "/user/login" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "Logs user into the system",
        "operationId" : "loginUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "creds",
          "in" : "query",
          "description" : "The string with user credentials in base64",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful login",
            "headers" : {
              "X-Expires-After" : {
                "type" : "string",
                "format" : "date-time",
                "description" : "Date in UTC when token expires"
              }
            },
            "schema" : {
              "$ref" : "#/definitions/Authorization"
            }
          },
          "400" : {
            "description" : "Invalid login credentials supplied"
          }
        },
        "security" : [ ]
      }
    },
    "/user/logout" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "Logs out current logged in user session",
        "operationId" : "logoutUser",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "creds",
          "in" : "query",
          "description" : "The string with user credentials in base64",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "Successfully logged off"
          }
        }
      }
    },
...
    "/studentIdentityCard/{userId}" : {
      "get" : {
        "tags" : [ "StudentIdentityCard" ],
        "summary" : "Get Student Identity Card for the User",
        "operationId" : "getStudentIdentityCard",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Student Identity Card model",
            "schema" : {
              "$ref" : "#/definitions/StudentIdentityCard"
            }
          },
          "404" : {
            "description" : "Student Identity Card for this {userId} not found"
          }
        }
      }
    },
    "/studentIdentityCard/{userId}/image/small" : {
      "get" : {
        "tags" : [ "StudentIdentityCard" ],
        "summary" : "Get small profile image for the Student Identity Card which belongs to the User",
        "operationId" : "getStudentIdentityCardImageSmall",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Student Identity Card small profile image"
          },
          "404" : {
            "description" : "Student Identity Card small image for this {userId} not found"
          }
        }
      }
    },
    "/studentIdentityCard/{userId}/signature" : {
      "get" : {
        "tags" : [ "StudentIdentityCard" ],
        "summary" : "Get signature image for the Student Identity Card which belongs to the User",
        "operationId" : "getStudentIdentityCardSignature",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Student Identity Card signature image"
          },
          "404" : {
            "description" : "Student Identity Card signature image for this {userId} not found"
          }
        }
      }
    },
    "/studentIdentityCard/{userId}/qr" : {
      "get" : {
        "tags" : [ "StudentIdentityCard" ],
        "summary" : "Get QR code for the Student Identity Card which belongs to the User",
        "operationId" : "getStudentIdentityCardQr",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Student Identity Card QR code"
          },
          "404" : {
            "description" : "Student Identity Card QR code for this {userId} not found"
          }
        }
      }
    },
    "/studentIdentityCard/{userId}/nfc" : {
      "get" : {
        "tags" : [ "StudentIdentityCard" ],
        "summary" : "Get NFC code for the Student Identity Card which belongs to the User",
        "operationId" : "getStudentIdentityCardNfc",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Student Identity Card NFC code"
          },
          "404" : {
            "description" : "Student Identity Card NFC code for this {userId} not found"
          }
        }
      }
    },
    "/studentIdentityCard/{userId}/image" : {
      "get" : {
        "tags" : [ "StudentIdentityCard" ],
        "summary" : "Get profile image for the Student Identity Card which belongs to the User",
        "operationId" : "getStudentIdentityCardImage",
        "produces" : [ "image/jpeg" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Student Identity Card profile image"
          },
          "404" : {
            "description" : "Student Identity Card image for this {userId} not found"
          }
        }
      },
      "put" : {
        "tags" : [ "StudentIdentityCard" ],
        "summary" : "Update profile image for the Student Identity Card which belongs to the User",
        "operationId" : "putStudentIdentityCardImage",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "image",
          "in" : "formData",
          "description" : "Image to upload",
          "required" : true,
          "type" : "file"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully uploaded Student Identity Card profile image"
          },
          "400" : {
            "description" : "Provided image is invalid"
          }
        }
      }
    },
    "/studentIdentityCard/{userId}/validate" : {
      "put" : {
        "tags" : [ "StudentIdentityCard" ],
        "summary" : "Request to revalidate and extend expiration for Student Identity Card",
        "operationId" : "putStudentIdentityCardValidation",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully requested Student Identity Card revalidation"
          },
          "404" : {
            "description" : "Passport for this {userId} not found"
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "Token" : {
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "StudentIdentityCard" : {
      "type" : "object",
      "properties" : {
        "studentId" : {
          "type" : "string",
          "example" : "53675"
        },
        "name" : {
          "type" : "string",
          "example" : "Unė"
        },
        "surname" : {
          "type" : "string",
          "example" : "Universitaitė"
        },
        "personalNo" : {
          "type" : "integer",
          "example" : 48703030001
        },
        "studiesAt" : {
          "type" : "string",
          "example" : "Socialinių mokslų f."
        },
        "validFrom" : {
          "type" : "string",
          "format" : "yyyy-mm-dd",
          "example" : "2017-09-01"
        },
        "validUntil" : {
          "type" : "string",
          "format" : "yyyy-mm-dd",
          "example" : "2021-06-30"
        },
        "spd" : {
          "type" : "integer",
          "example" : 701303
        },
        "barcode" : {
          "type" : "string",
          "example" : "94403905000000014"
        }
      },
      "example" : {
        "studentId" : "53675",
        "studiesAt" : "Socialinių mokslų f.",
        "surname" : "Universitaitė",
        "spd" : 701303,
        "name" : "Unė",
        "personalNo" : 48703030001,
        "validUntil" : "2021-06-30",
        "validFrom" : "2017-09-01",
        "barcode" : "94403905000000014"
      }
    },
...
    "Authorization" : {
      "type" : "object",
      "properties" : {
        "userId" : {
          "type" : "string",
          "example" : "31234567890"
        },
        "token" : {
          "type" : "string",
          "example" : "Basic ZllvSFpwcnJCR01XZXExZk9IYUFIWTZmWE53bTA1TDJLaXdZRFI2azZrWDYyclhTY3REOEhVbHZVRzZGcEVlOA=="
        }
      },
      "example" : {
        "userId" : "31234567890",
        "token" : "Basic ZllvSFpwcnJCR01XZXExZk9IYUFIWTZmWE53bTA1TDJLaXdZRFI2azZrWDYyclhTY3REOEhVbHZVRzZGcEVlOA=="
      }
    }
  }
}